refactor: move gorm setup functions
